-ampoule allumée (jaune (plutôt orange)avec trait autour) - panneau en rouge
Home page :
-activate link to the tools
-ressources : enlever  https
WHO report, Safe Future Methodology, Best Interest of the Child UNHCR report, CBSS handbook, CompAct study, Prosecuting THB  Eurojust report

-insert popup “benefits of integration measures”

"Depending on the country of destination, integration measures may include: access to accommodation, languages courses, vocational training, psychosocial and health services, legal assistance, etc.
To note: access to specific and/or mainstream integration measures often depends on the victim’s status in the country of destination (different types of residence permit). It is important to advocate for the granting of long-term residence permit for victims of trafficking, as it is generally the main ground for integration to take place.”

—————
- Amend the pop-up “travel to country of origin” as follow (amendments in red):
Travel arrangement includes:
- Acquisition of travel documents
- Purchase of air/train ticket to the final destination
The role of embassies and consulates is key to get travel document, their staff should thus be sensitized and trained on the identification of victims of trafficking and referral.
- ( withdraw bullet point) Ensure safety during travel: direct flights should be privileged, assistance in transit provided if need be, operational escort (compulsory for children VoT) and/or medical escorts organized.
—————

- In the pop-up “Established”, please add a “a” to “repeating”

—————

- Shorten the last box of the flowchart (“Counselling and monitoring (…) independent”) as the white box is too big compared to the text size.

—————

- The EU banner is too large, please reduce it.  You can reduce the text police to size 14 and align the EU flag on the left of the sentence. This comment applies for all the pages where the EU banner is inserted.

TRM page :

-insert new text content (see in Content January)
-reduce banner

TACT project page :

-justify text
-Replace Font Gill san MT
-center title
-delete TACT logo at the end
-add sentence :  “For more information on the TACT project, please contact:”
-change sentence It has been funded by the EU Asylum, Migration and Integration Fund (AMIF).
-reduce banner